+++ crates/tuirealm_derive/CHANGELOG.md
@@ -1,10 +1,19 @@
 # Changelog
 
 - [Changelog](#changelog)
+  - [2.0.1](#201)
   - [2.0.0](#200)
   - [1.0.0](#100)
 
 ---
+
+## 2.0.1
+
+Released on 08/06/2025
+
+- Fix: Added support for structs with lifetime and generics
+- It is now possible to specify the component to apply the `#derive(MockComponent)` to, using the `#[component("ComponentName")]` attribute.
+- **MSRV**: 1.64
 
 ## 2.0.0